Added example policy to automate K10 policy creation based on SLA presets
New policy to enforce setting a dataprotection label on a newly created namespace, corresponding to a K10 policy preset, which will then be used to generate a policy based on the centrally configured preset settings.

Generate policy placed in the same directory as the two policies are designed to be used together.