Added Kasten PolicyPreset-based data protection guardrail policies
New example policy to enforce setting a dataprotection label on a newly created namespace, corresponding to a K10 policy preset, which will then be used to generate a K10 backup policy based on the centrally configured preset settings.

Generate policy placed in the same directory as the two policies are designed to be used together.

Signed-off-by: Matt Bator <matt.bator@kasten.io>