kubevirt: Only enforce use of instance types and preferences
The policy would previously check for the use of `spec.template.domain`
that is not actually invalid when using instance types and preferences.

This change drops that check and replaces it with two validations
ensuring both `spec.instancetype.name` and `spec.preference.name` are
provided. Any remaining validation being left to the submission webhooks
within KubeVirt itself.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>